# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafe Backend API",
        "description": "Dokumentasi API Cafe Backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/guest/order": {
            "post": {
                "tags": [
                    "Guest Orders"
                ],
                "summary": "Membuat order baru",
                "description": "Menerima data order dari guest. Bisa berupa cart_json (dari Blade FE) atau array products langsung.",
                "operationId": "98d7a63b20bf04f779f35b5ef0eac1a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table",
                                    "customer_name",
                                    "customer_phone",
                                    "payment_method",
                                    "products"
                                ],
                                "properties": {
                                    "table": {
                                        "description": "Nomor/meja pelanggan",
                                        "type": "string",
                                        "example": "A1"
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "Budi"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "cash"
                                    },
                                    "products": {
                                        "description": "Daftar produk yang dipesan",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "cart_json": {
                                        "description": "Alternatif input jika request berasal dari Blade FE",
                                        "type": "string",
                                        "example": "[{\"id\":1,\"quantity\":2},{\"id\":3,\"quantity\":1}]",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        },
                                        "data": {
                                            "description": "Detail order beserta transaksi",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "table": {
                                                    "type": "string",
                                                    "example": "A1"
                                                },
                                                "customer_name": {
                                                    "type": "string",
                                                    "example": "Budi"
                                                },
                                                "customer_phone": {
                                                    "type": "string",
                                                    "example": "08123456789"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 75000
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "product_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "product_name": {
                                                                "type": "string",
                                                                "example": "Cappuccino"
                                                            },
                                                            "category_name": {
                                                                "type": "string",
                                                                "example": "Minuman"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 25000
                                                            },
                                                            "total": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 50000
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "transaction": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "payment_method": {
                                                            "type": "string",
                                                            "example": "cash"
                                                        },
                                                        "total": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 75000
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT dengan format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}